chore(github-actions): Update GitHub Actions versions in Docker workflow
Some of the actions versions that are used in the in Docker workflow are outdated (e.g. `docker/setup-buildx-action@v2.5.0`), the versions were updated to use the major version tags (e.g. `docker/setup-buildx-action@v2`) as it's done for other actions such as `actions/checkout`. This way the latest minor and patch versions can be used when the workflow runs, and the amount of times that the version needs to be updated in the workflow files is reduced.